Fixed some errors and altered Opcode section a bit
For the MOVE portion of effective addressing (opcode deciphering), it checks whether the first 4 bits (bit endian) make up 1, 2, or 3. Then it checks whether the opcode is MOVEA or just MOVE. The size is stored into D3 to compare in the OPCODES section, where it can attach '.B', '.W', or '.L' to 'MOVE', and '.W', or '.L' to 'MOVEA'.